Create Code.py
The code. Please make a pull request if there is a more efficient way to do this above process or if you find a fault in the logic of the program. I would be very grateful and you would be doing a good thing. I am a newbie to all this and this is just my second program. If you are in a position where you can share your knowledge for the greater good, please do so.